backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
django-cors-headers==4.8.0
Pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
